json viewer dark mode and select menu md view
